sinks/interamap/requirements.txt
geocoder
folium
fastkml
PySide6
offline_folium
PyQt6
qrcode
pillow